Python/Automated News Website/requirements.txt
ago==0.0.93
alembic==1.0.5
amqp==2.4.0
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
awscli==1.16.68
beautifulsoup4==4.6.3
billiard==3.5.0.5
botocore==1.12.58
breadability==0.1.20
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
constantly==15.1.0
croniter==0.3.26
cryptography==2.4.2
cssselect==1.0.3
docopt==0.6.2
docutils==0.14
dotmap==1.3.4
elasticsearch==6.3.1
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
Flask-CLI==0.4.0
Flask-Migrate==2.3.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
hjson==3.0.1
hurry.filesize==0.9
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10
jmespath==0.9.3
kombu==4.2.2.post1
langdetect==1.0.7
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.1.0
news-please==1.2.44
newspaper3k==0.2.8
nltk==3.4
numpy==1.15.4
parsel==1.5.1
Pillow==5.3.0
pkg-resources==0.0.0
plac==1.0.0
psutil==5.4.8
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.9
PyYAML==3.13
queuelib==1.5.0
readability-lxml==0.7
redis==3.0.1
requests==2.17.1
requests-file==1.4.3
rsa==3.4.2
s3transfer==0.1.13
Scrapy==1.5.1
service-identity==17.0.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.14
sumy==0.7.0
tinysegmenter==0.3
tldextract==2.2.0
Twisted==18.9.0
tzlocal==1.5.1
urllib3==1.21.1
vine==1.2.0
w3lib==1.19.0
warcio==1.6.3
Werkzeug==0.14.1
WTForms==2.2.1
zope.interface==4.6.0
